+++ test/ethers/owner_contract_test.exs
@@ -13,8 +13,11 @@ defmodule Ethers.OwnerContractTest do
   @sample_address "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
 
   test "can deploy and get owner" do
-    init_params = OwnerContract.constructor(@sample_address)
-    assert {:ok, tx_hash} = Ethers.deploy(OwnerContract, init_params, %{from: @from})
+    encoded_constructor = OwnerContract.constructor(@sample_address)
+
+    assert {:ok, tx_hash} =
+             Ethers.deploy(OwnerContract, encoded_constructor: encoded_constructor, from: @from)
+
     assert {:ok, address} = Ethers.deployed_address(tx_hash)
 
     assert {:ok, @sample_address} = OwnerContract.get_owner() |> Ethers.call(to: address)
@@ -22,17 +25,22 @@ defmodule Ethers.OwnerContractTest do
 
   describe "overriding RPC options" do
     test "can override RPC client" do
-      init_params = OwnerContract.constructor(@sample_address)
+      encoded_constructor = OwnerContract.constructor(@sample_address)
 
       assert {:ok, "tx_hash"} =
-               Ethers.deploy(OwnerContract, init_params, %{from: @from},
+               Ethers.deploy(OwnerContract,
+                 encoded_constructor: encoded_constructor,
+                 from: @from,
                  rpc_client: Ethers.TestRPCModule
                )
     end
 
     test "can override RPC options" do
-      init_params = OwnerContract.constructor(@sample_address)
-      assert {:ok, tx_hash} = Ethers.deploy(OwnerContract, init_params, %{from: @from})
+      encoded_constructor = OwnerContract.constructor(@sample_address)
+
+      assert {:ok, tx_hash} =
+               Ethers.deploy(OwnerContract, encoded_constructor: encoded_constructor, from: @from)
+
       assert {:ok, address} = Ethers.deployed_address(tx_hash)
 
       assert {:ok, @sample_address} =