+++ test/ethers/registry_contract_test.exs
@@ -83,8 +83,14 @@ defmodule Ethers.RegistryContractTest do
   end
 
   defp deploy_registry_contract(_ctx) do
-    init_params = RegistryContract.constructor()
-    assert {:ok, tx_hash} = Ethers.deploy(RegistryContract, init_params, %{from: @from})
+    encoded_constructor = RegistryContract.constructor()
+
+    assert {:ok, tx_hash} =
+             Ethers.deploy(RegistryContract,
+               encoded_constructor: encoded_constructor,
+               from: @from
+             )
+
     assert {:ok, address} = Ethers.deployed_address(tx_hash)
 
     [address: address]